Inline OrbitalMotion methods into orbitalMotion.hpp
